+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Option `font.builtin_box_drawing` to disable the built-in font for drawing box characters
 - Track and report surface damage information to Wayland compositors
-- Escape sequence for undercurl (`CSI 4 : 3 m`)
+- Escape sequence for undercurl, dotted and dashed underlines (`CSI 4 : [3-5] m`)
 
 ### Changed